+++ src/main/java/elbar/company/learn_center_rest/service/system/skill/SkillServiceImpl.java
@@ -18,9 +18,11 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
+import org.webjars.NotFoundException;
 
 import java.time.LocalDateTime;
 import java.util.List;
+import java.util.Optional;
 import java.util.UUID;
 
 @Service
@@ -39,7 +41,11 @@ public class SkillServiceImpl extends AbstractService<SkillValidator, SkillMappe
     @Override
     public ResponseEntity<Data<Void>> update(SkillUpdateDTO DTO) {
         validator.validOnUpdate(DTO);
-        Skill skill = repository.getByCode(DTO.getCode());
+        Optional<Skill> optional = repository.getByCode(DTO.getCode());
+        if (optional.isEmpty()) {
+            throw new NotFoundException("Skill not found");
+        }
+        Skill skill = optional.get();
         skill.setName(DTO.getName());
         skill.setPublished(DTO.isPublished());
         skill.setUpdatedAt(LocalDateTime.now());
@@ -50,6 +56,10 @@ public class SkillServiceImpl extends AbstractService<SkillValidator, SkillMappe
     @Override
     public ResponseEntity<Data<Void>> delete(UUID key) {
         validator.validateKey(key);
+        Optional<Skill> optional = repository.getByCode(key);
+        if (optional.isEmpty()) {
+            throw new NotFoundException("Skill not found");
+        }
         repository.deleteByCode(key);
         return new ResponseEntity<>(new Data<>(true), HttpStatus.OK);
     }
@@ -57,19 +67,18 @@ public class SkillServiceImpl extends AbstractService<SkillValidator, SkillMappe
     @Override
     public ResponseEntity<Data<SkillGetDTO>> get(UUID key) {
         validator.validateKey(key);
-        return new ResponseEntity<>(new Data<>(mapper.fromGetDTO(repository.getByCode(key))), HttpStatus.OK);
+        return new ResponseEntity<>(new Data<>(mapper.fromGetDTO(repository.getByCode(key).orElseThrow(() -> new NotFoundException("Skill not found")))), HttpStatus.OK);
     }
 
     @Override
     public ResponseEntity<Data<SkillDetailDTO>> detail(UUID key) {
         validator.validateKey(key);
-        return new ResponseEntity<>(new Data<>(mapper.fromDetailDTO(repository.getByCode(key))), HttpStatus.OK);
+        return new ResponseEntity<>(new Data<>(mapper.fromDetailDTO(repository.getByCode(key).orElseThrow(() -> new NotFoundException("Skill not found")))), HttpStatus.OK);
     }
 
     @Override
     public ResponseEntity<Data<List<SkillGetDTO>>> list(SkillCriteria criteria) {
-        PageRequest request = PageRequest.of(criteria.getPage(), criteria.getSize());
-        Page<Skill> all = repository.findAll(request);
-        return new ResponseEntity<>(new Data<>(mapper.fromGetListDTO(all.toList()), all.getSize()), HttpStatus.OK);
+        PageRequest request = PageRequest.of(criteria.getPage(), criteria.getSize(), criteria.getSort(), criteria.getFieldsEnum().getValue());
+        return new ResponseEntity<>(new Data<>(mapper.fromGetListDTO(repository.findAll(request).stream().toList()), repository.count()), HttpStatus.OK);
     }
 }
