+++ src/main/java/elbar/company/learn_center_rest/service/system/tag/TagServiceImpl.java
@@ -16,9 +16,11 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
+import org.webjars.NotFoundException;
 
 import java.time.LocalDateTime;
 import java.util.List;
+import java.util.Optional;
 import java.util.UUID;
 
 @Service
@@ -37,7 +39,11 @@ public class TagServiceImpl extends AbstractService<TagValidator, TagMapper, Tag
     @Override
     public ResponseEntity<Data<Void>> update(TagUpdateDTO DTO) {
         validator.validOnUpdate(DTO);
-        Tag tag = repository.getByCode(DTO.getCode());
+        Optional<Tag> optional = repository.getByCode(DTO.getCode());
+        if (optional.isEmpty()) {
+            throw new NotFoundException("Tag not found");
+        }
+        Tag tag = optional.get();
         tag.setName(DTO.getName());
         tag.setPublished(DTO.isPublished());
         tag.setUpdatedAt(LocalDateTime.now());
@@ -48,6 +54,10 @@ public class TagServiceImpl extends AbstractService<TagValidator, TagMapper, Tag
     @Override
     public ResponseEntity<Data<Void>> delete(UUID key) {
         validator.validateKey(key);
+        Optional<Tag> optional = repository.getByCode(key);
+        if (optional.isEmpty()) {
+            throw new NotFoundException("Tag not found");
+        }
         repository.deleteByCode(key);
         return new ResponseEntity<>(new Data<>(true), HttpStatus.OK);
     }
@@ -55,19 +65,18 @@ public class TagServiceImpl extends AbstractService<TagValidator, TagMapper, Tag
     @Override
     public ResponseEntity<Data<TagGetDTO>> get(UUID key) {
         validator.validateKey(key);
-        return new ResponseEntity<>(new Data<>(mapper.fromGetDTO(repository.getByCode(key))), HttpStatus.OK);
+        return new ResponseEntity<>(new Data<>(mapper.fromGetDTO(repository.getByCode(key).orElseThrow(() -> new NotFoundException("Tag not found")))), HttpStatus.OK);
     }
 
     @Override
     public ResponseEntity<Data<TagDetailDTO>> detail(UUID key) {
         validator.validateKey(key);
-        return new ResponseEntity<>(new Data<>(mapper.fromDetailDTO(repository.getByCode(key))), HttpStatus.OK);
+        return new ResponseEntity<>(new Data<>(mapper.fromDetailDTO(repository.getByCode(key).orElseThrow(() -> new NotFoundException("Tag not found")))), HttpStatus.OK);
     }
 
     @Override
     public ResponseEntity<Data<List<TagGetDTO>>> list(TagCriteria criteria) {
-        PageRequest request = PageRequest.of(criteria.getPage(), criteria.getSize());
-        Page<Tag> all = repository.findAll(request);
-        return new ResponseEntity<>(new Data<>(mapper.fromGetListDTO(all.toList()), all.getSize()), HttpStatus.OK);
+        PageRequest request = PageRequest.of(criteria.getPage(), criteria.getSize(), criteria.getSort(), criteria.getFieldsEnum().getValue());
+        return new ResponseEntity<>(new Data<>(mapper.fromGetListDTO(repository.findAll(request).stream().toList()), repository.count()), HttpStatus.OK);
     }
 }
