+++ src/main/java/elbar/company/learn_center_rest/service/system/transaction_status/TransactionStatusServiceImpl.java
@@ -16,9 +16,11 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
+import org.webjars.NotFoundException;
 
 import java.time.LocalDateTime;
 import java.util.List;
+import java.util.Optional;
 import java.util.UUID;
 
 @Service
@@ -37,7 +39,11 @@ public class TransactionStatusServiceImpl extends AbstractService<TransactionSta
     @Override
     public ResponseEntity<Data<Void>> update(TransactionStatusUpdateDTO DTO) {
         validator.validOnUpdate(DTO);
-        TransactionStatus status = repository.getByCode(DTO.getCode());
+        Optional<TransactionStatus> optional = repository.getByCode(DTO.getCode());
+        if (optional.isEmpty()) {
+            throw new NotFoundException("Transaction Status not found");
+        }
+        TransactionStatus status = optional.get();
         status.setName(DTO.getName());
         status.setPublished(DTO.isPublished());
         status.setUpdatedAt(LocalDateTime.now());
@@ -48,6 +54,10 @@ public class TransactionStatusServiceImpl extends AbstractService<TransactionSta
     @Override
     public ResponseEntity<Data<Void>> delete(UUID key) {
         validator.validateKey(key);
+        Optional<TransactionStatus> optional = repository.getByCode(key);
+        if (optional.isEmpty()) {
+            throw new NotFoundException("Transaction Status not found");
+        }
         repository.deleteByCode(key);
         return new ResponseEntity<>(new Data<>(true), HttpStatus.OK);
     }
@@ -55,19 +65,18 @@ public class TransactionStatusServiceImpl extends AbstractService<TransactionSta
     @Override
     public ResponseEntity<Data<TransactionStatusGetDTO>> get(UUID key) {
         validator.validateKey(key);
-        return new ResponseEntity<>(new Data<>(mapper.fromGetDTO(repository.getByCode(key))), HttpStatus.OK);
+        return new ResponseEntity<>(new Data<>(mapper.fromGetDTO(repository.getByCode(key).orElseThrow(() -> new NotFoundException("Transaction Status not found")))), HttpStatus.OK);
     }
 
     @Override
     public ResponseEntity<Data<TransactionStatusDetailDTO>> detail(UUID key) {
         validator.validateKey(key);
-        return new ResponseEntity<>(new Data<>(mapper.fromDetailDTO(repository.getByCode(key))), HttpStatus.OK);
+        return new ResponseEntity<>(new Data<>(mapper.fromDetailDTO(repository.getByCode(key).orElseThrow(() -> new NotFoundException("Transaction Status not found")))), HttpStatus.OK);
     }
 
     @Override
     public ResponseEntity<Data<List<TransactionStatusGetDTO>>> list(TransactionStatusCriteria criteria) {
-        PageRequest request = PageRequest.of(criteria.getPage(), criteria.getSize());
-        Page<TransactionStatus> all = repository.findAll(request);
-        return new ResponseEntity<>(new Data<>(mapper.fromGetListDTO(all.toList()), all.getSize()), HttpStatus.OK);
+        PageRequest request = PageRequest.of(criteria.getPage(), criteria.getSize(), criteria.getSort(), criteria.getFieldsEnum().getValue());
+        return new ResponseEntity<>(new Data<>(mapper.fromGetListDTO(repository.findAll(request).stream().toList()), repository.count()), HttpStatus.OK);
     }
 }
